Version 1.0-RC4
Fixed BSP UART driver to read 9-bit data or data with parity bit from RS-232 channels
Added workaround into BSP UART IRQ handler if UART error occurred before DMA receiver is enabled
Fixed display of LIN support in case of not presettings
Added entry to upper menu "Algorithm" from menu item "Defaults->YES"
Generated documentation 1.1
Extended script for UART sending
Added note for deprecated algorithm in MATLAB

+++ project/application/inc/cli.h
@@ -64,7 +64,7 @@ void cli_trace(const char *format, ...);
  */
 uint8_t cli_rs232_trace(enum uart_type uart_type,
                         enum rs232_trace_type trace_type,
-                        uint8_t *data,
+                        uint16_t *data,
                         uint32_t len,
                         bool break_line);
